pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "prospectivity-tools"
version = "0.2.0"
description = "Geospatial prospectivity heatmap using H3 indexing"
readme = "README.md"
requires-python = "==3.10.*"

dependencies = [
  "geopandas==0.12.0",
  "shapely==1.8.5",
  "pyogrio==0.4.1",
  "h3==3.7.6",
  "folium==0.12.1",
  "pydantic==1.10.2",
  "click==8.1",
  "pyarrow==10.0.1",
  "pyproj==3.4.1",
  "numpy==1.23.5",
  "pandas==1.5.3",
  "ipykernel>=6.20.0",
  "pyyaml==6.0",
  "mapclassify==2.4.3",
  "matplotlib==3.5.3",
  "llama-index>=0.8.69",
  "langchain>=0.0.354",
  "contextily>=1.6.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5",
  "ruff>=0.4.0",
  "black>=24.7",
  "pre-commit>=3.6"
]

[project.scripts]
prospectivity = "prospectivity_tools.cli:main"

[tool.black]
target-version = ["py310"]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py310"
lint.select = ["E", "F", "B", "I", "UP"]

[tool.uv]
package = true
